Airflow_Codes/airflow_setup/requirements.txt
apache-airflow-providers-google
apache-airflow-providers-postgres
apache-airflow-providers-http
apache-airflow-providers-dbt-cloud
google-cloud-storage
pyarrow
sodapy